move tests into test folder
