+++ template/web/src/hooks/test/useMediaQuery.test.ts
@@ -3,7 +3,7 @@
  */
 
 import { renderHook } from '@testing-library/react';
-import { useMediaQuery } from './useMediaQuery';
+import { useMediaQuery } from '../useMediaQuery';
 
 function mockMatchMedia(matches: boolean) {
   Object.defineProperty(window, 'matchMedia', {
