+++ template/web/src/utils/test/address.test.ts
@@ -1,4 +1,4 @@
-import { getSlicedAddress } from './address';
+import { getSlicedAddress } from '../address';
 
 describe('getSlicedAddress', () => {
   it('should return a string of class names', () => {
