+++ template/web/src/utils/test/balance.test.ts
@@ -1,4 +1,4 @@
-import { getAccountBalance } from './balance';
+import { getAccountBalance } from '../balance';
 
 describe('getAccountBalance', () => {
   it('should return a string of class names', () => {
