+++ template/web/src/utils/test/ipfsToHTTP.test.ts
@@ -1,4 +1,4 @@
-import { ipfsToHTTP } from './ipfs';
+import { ipfsToHTTP } from '../ipfs';
 
 describe('ipfsToHTTP', () => {
   it('converts ipfs URI to HTTPS URI', () => {
